png: Initial encoder.
This includes some restructuring to better arrange code for  decoders and encoders, and a simple working encoder for png with some filtering support
